State: Try typing notices actions
