--- client/state/notices/actions.js
@@ -1,43 +1,0 @@
-/**
- * External dependencies
- */
-import impureLodash from 'calypso/lib/impure-lodash';
-
-/**
- * Internal dependencies
- */
-import { NOTICE_CREATE, NOTICE_REMOVE } from 'calypso/state/action-types';
-import { extendAction } from 'calypso/state/utils';
-
-import 'calypso/state/notices/init';
-
-const { uniqueId } = impureLodash;
-
-export function removeNotice( noticeId ) {
-	return {
-		noticeId: noticeId,
-		type: NOTICE_REMOVE,
-	};
-}
-
-export function createNotice( status, text, { id, ...noticeOptions } = {} ) {
-	return {
-		type: NOTICE_CREATE,
-		notice: Object.assign( { showDismiss: true }, noticeOptions, {
-			noticeId: id || uniqueId(),
-			status,
-			text,
-		} ),
-	};
-}
-
-export const successNotice = createNotice.bind( null, 'is-success' );
-export const errorNotice = createNotice.bind( null, 'is-error' );
-export const infoNotice = createNotice.bind( null, 'is-info' );
-export const warningNotice = createNotice.bind( null, 'is-warning' );
-export const plainNotice = createNotice.bind( null, 'is-plain' );
-
-// Higher-order action creator: modify the wrapped creator to return actions with the
-// `notices.skip` meta, so that it's ignored by the notices middleware.
-export const withoutNotice = ( actionCreator ) => ( ...args ) =>
-	extendAction( actionCreator( ...args ), { meta: { notices: { skip: true } } } );
